// src/modules/contact/contact.interface.ts
export interface IContactInfo {
	id?: string;
	heroTitle?: string;
	heroDescription?: string;
	address?: string;
	phone?: string;
	whatsapp?: string;
	email?: string;
	secondaryEmail?: string;
	website?: string;
	businessHours?: string;
	mapEmbedUrl?: string;
	siteVisitNotice?: string;
	imageUrl?: string;
	facebookUrl?: string;
	youtubeUrl?: string;
}